Moved alignment and max_length to a separate visitor
