+++ testmain.py
@@ -2,6 +2,7 @@ from antlr4 import *
 from JavaLexer import JavaLexer
 from JavaParser import JavaParser
 from FormattingVisitor import FormattingVisitor
+from AlignmentVisitor import AlignmentVisitor
 from ErrorLogger import ErrorLogger
 from ConfigClass import ConfigClass
 import re
@@ -25,9 +26,21 @@ def parse_java_code(file_path):
     parser = JavaParser(tokens)
     tree = parser.compilationUnit()
 
-    #print(tree.toStringTree(recog=parser))
-
     return tree, tokens
+
+def format_code(tree, tokens, configs):
+    formatter = FormattingVisitor(tokens, configs)
+    first_code_pass = formatter.get_formatted_code(tree)
+    
+    lexer = JavaLexer(InputStream(first_code_pass))
+    tokens = CommonTokenStream(lexer)
+    parser = JavaParser(tokens)
+    tree = parser.compilationUnit()
+
+    aligner = AlignmentVisitor(tokens, configs)
+    second_code_pass = aligner.get_formatted_code(tree)
+
+    return second_code_pass
 
 tree, tokens = parse_java_code("test2.java")
 configs = load_config(".java-format.json")
@@ -40,5 +53,5 @@ if errors:
     for error in errors:
         print(error)
 
-formatted_code = formatter.get_formatted_code(tree)
+formatted_code = format_code(tree, tokens, configs)
 print(formatted_code)
